requirements.txt
Django
Pillow
django-multiselectfield
asgiref
sqlparse,
django-tinymce
django-unfold